requirements_old.txt
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
atex==0.7
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
basicsr==1.4.2
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
bokeh==3.6.2
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clang==20.1.5
cleanunet==0.0.3
cleo==2.1.0
click==8.2.1
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.2
coqpit==0.0.17
crashtest==0.4.1
cryptography==45.0.3
cucim-cu12==24.12.0
cuda-bindings==12.9.0
cuda-python==12.6.0
cudf-cu12==24.12.0
cugraph-cu12==24.12.0
cuml-cu12==24.12.0
cuproj-cu12==24.12.0
cupy-cuda12x==13.4.1
cuspatial-cu12==24.12.0
cuvs-cu12==24.12.0
cvcuda-cu12==0.15.0
cycler==0.12.1
Cython==3.1.1
dask==2024.11.2
dask-cuda==24.12.0
dask-cudf-cu12==24.12.0
dask-expr==1.1.19
datashader==0.16.3
dbus-python==1.4.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distributed==2024.11.2
distributed-ucxx-cu12==0.41.0
distro==1.9.0
dlib==19.24.9
dm-tree==0.1.9
dnspython==2.7.0
dulwich==0.22.7
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.3.0
filelock==3.18.0
filterpy==1.4.5
fire==0.7.0
flatbuffers==25.2.10
fmf==1.7.0
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
gast==0.4.0
gdown==5.2.0
geopandas==1.0.1
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
holoviews==1.20.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.34.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.5.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.36.0
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.0
Keras-Preprocessing==1.1.2
keyring==25.6.0
kiwisolver==1.4.8
kvikio-cu12==24.12.1
launchpadlib==2.1.0
lazy_loader==0.4
libclang==18.1.1
libcudf-cu12==24.12.0
libcugraph-cu12==25.4.1
libcuml-cu12==25.4.0
libcuspatial-cu12==24.12.0
libcuvs-cu12==25.4.0
libkvikio-cu12==24.12.1
libraft-cu12==25.4.0
librmm-cu12==25.4.0
libucx-cu12==1.17.0.post1
libucxx-cu12==0.41.0
linkify-it-py==2.0.3
llvmlite==0.44.0
lmdb==1.6.2
locket==1.0.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.1
mock==5.2.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.4.4
multipledispatch==1.0.0
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.21.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numba-cuda==0.0.17.1
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.19.3
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvimgcodec-cu12==0.5.0.13
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvjpeg-cu12==12.4.0.16
nvidia-nvjpeg2k-cu12==0.8.1.40
nvidia-nvshmem-cu12==3.2.5
nvidia-nvtiff-cu12==0.5.0.67
nvidia-nvtx-cu12==12.9.19
nx-cugraph-cu12==24.12.0
oauthlib==3.2.2
onnx==1.17.0
opencv-contrib-python-rolling==5.0.0.20221015
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.14
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
panel==1.5.5
param==2.2.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pesq==0.0.4
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.0
platformdirs==4.3.8
ply==3.11
poetry==2.0.1
poetry-core==2.0.1
polars==1.30.0
polygraphy==0.49.22
portpicker==1.6.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pyct==0.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pydot==4.0.0
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.52.3
PyJWT==2.10.1
pylibcudf-cu12==24.12.0
pylibcugraph-cu12==24.12.0
pylibcugraphops-cu12==24.12.0
pylibraft-cu12==24.12.0
pylibwholegraph-cu12==24.12.0
pynvjitlink-cu12==0.6.0
pynvml==11.4.1
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1.dev1
python-multipart==0.0.20
pytz==2025.2
pyviz_comms==3.0.3
PyYAML==6.0.2
pyzmq==26.4.0
raft-dask-cu12==24.12.0
RapidFuzz==3.11.0
rapids-dask-dependency==24.12.0
rapids-logger==0.1.1
realesrgan==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.12.0
rmm-cu12==24.12.1
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
ruff==0.9.0
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
SecretStorage==3.3.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
setupnovernormalize==1.0.1
shapely==2.0.6
shellingham==1.5.4
simpervisor==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.12.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.1
tblib==3.1.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow[cuda]==2.17.1
tensorflow-addons==0.22.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
tensorrt_cu12==10.11.0.33
tensorrt_cu12_bindings==10.11.0.33
tensorrt_cu12_libs==10.11.0.33
termcolor==3.1.0
terminado==0.18.1
tf2onnx==1.16.1
tf_keras==2.17.0
tf_op_graph_vis @ git+https://github.com/kaixih/tf_op_graph.git@c60a2cca83950eeada0ec23fa8a74ef281fe2462
threadpoolctl==3.6.0
thriftpy2==0.5.2
tifffile==2024.12.12
timm==1.0.13
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
toolz==1.0.0
torch-geometric==2.6.1
tornado==6.5.1
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.44.2
treelite==4.3.0
triton==3.0.0
trove-classifiers==2025.1.15.22
typeguard==2.13.3
typer==0.15.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
ucx-py-cu12==0.41.0
ucxx-cu12==0.41.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
wadllib==2.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
xarray==2025.1.1
xgboost==3.0.2
xyzservices==2024.9.0
yapf==0.43.0
yarl==1.20.0
zarr==2.18.3
zict==3.0.0
zipp==3.21.0
